functions/requirements.txt
Flask==3.0.0
firebase-functions
firebase-admin
google-generativeai
Flask-Cors